// src/compute/types.generated.ts
export interface CalculatorInputsById {
// ...
  "chadsvasc": {
    "age_group": "under_65" | "age_65_74" | "age_75_plus";
    "sex": "male" | "female";
    "chf": boolean;
    "hypertension": boolean;
    "stroke_history": boolean;
    "vascular_disease": boolean;
    "diabetes": boolean;
  };
// ...
}

export interface CalculatorOutputsById {
// ...
  "chadsvasc": {
    "score": number;
    "stroke_risk": "0.2%" | "0.6%" | "2.2%" | "3.2%" | "4.8%" | "7.2%" | "9.7%" | "11.2%" | "10.8%" | "12.2%";
    "embolism_risk": "0.3%" | "0.9%" | "2.9%" | "4.6%" | "6.7%" | "10.0%" | "13.6%" | "15.7%" | "15.2%" | "17.4%";
    "stroke_risk_percent": number;
    "thromboembolism_risk_percent": number;
    "risk_time_horizon": "one_year";
  };
// ...
}